backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
pillow==10.1.0
pypdf2==3.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
